feat: 멀티 모듈 프로젝트 구성으류 변경
