+++ shilly-shally/core-domain/src/main/java/com/shillyshally/coredomain/restaurant/Restaurant.java
@@ -1,0 +1,40 @@
+package com.shillyshally.coredomain.restaurant;
+
+import jakarta.persistence.Entity;
+import jakarta.persistence.EnumType;
+import jakarta.persistence.Enumerated;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import lombok.AccessLevel;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Builder
+@Getter
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+@AllArgsConstructor
+@Entity
+public class Restaurant {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+
+    private String name;
+
+    @Enumerated(value = EnumType.STRING)
+    private Category category;
+
+    private String categoryDetail;
+
+    private String phone;
+
+    private String url;
+
+    private Double longitude;
+
+    private Double latitude;
+}
