+++ shilly-shally/external-api/build.gradle
@@ -1,0 +1,5 @@
+dependencies {
+    implementation 'org.springframework.boot:spring-boot-starter-web'
+    runtimeOnly 'com.h2database:h2'
+    implementation project(':core-domain')
+}
