+++ shilly-shally/external-api/src/main/java/com/shillyshally/externalapi/ExternalApiApplication.java
@@ -1,0 +1,17 @@
+package com.shillyshally.externalapi;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.boot.autoconfigure.domain.EntityScan;
+import org.springframework.data.jpa.repository.config.EnableJpaRepositories;
+
+@SpringBootApplication
+@EntityScan(basePackages = {"com.shillyshally.coredomain"})
+@EnableJpaRepositories(basePackages = {"com.shillyshally.coredomain"})
+public class ExternalApiApplication {
+
+    public static void main(String[] args) {
+        SpringApplication.run(ExternalApiApplication.class, args);
+    }
+
+}
