+++ shilly-shally/external-api/src/main/java/com/shillyshally/externalapi/application/dto/RestaurantResponse.java
@@ -1,0 +1,8 @@
+package com.shillyshally.externalapi.application.dto;
+
+public record RestaurantResponse(
+        long id,
+        String name
+) {
+
+}
