+++ shilly-shally/external-api/src/main/java/com/shillyshally/externalapi/config/WebConfig.java
@@ -1,0 +1,21 @@
+package com.shillyshally.externalapi.config;
+
+
+import org.springframework.context.annotation.Configuration;
+import org.springframework.web.servlet.config.annotation.CorsRegistry;
+import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
+
+@Configuration
+public class WebConfig implements WebMvcConfigurer {
+
+    private static final String ALLOWED_METHOD_NAMES = "GET,POST,PUT,DELETE,OPTIONS,PATCH";
+
+    @Override
+    public void addCorsMappings(final CorsRegistry registry) {
+        registry.addMapping("/**")
+                .allowedOrigins("*")
+                .allowedMethods(ALLOWED_METHOD_NAMES.split(","))
+                .exposedHeaders("*");
+    }
+}
+
