+++ shilly-shally/external-api/src/main/java/com/shillyshally/externalapi/exception/dto/ErrorResponse.java
@@ -1,0 +1,5 @@
+package com.shillyshally.externalapi.exception.dto;
+
+public record ErrorResponse(String message) {
+
+}
