+++ shilly-shally/internal-api/build.gradle
@@ -1,0 +1,3 @@
+dependencies {
+    implementation project(':core-domain')
+}
