+++ shilly-shally/internal-api/src/main/java/com/shillyshally/internalapi/InternalApiApplication.java
@@ -1,0 +1,13 @@
+package com.shillyshally.internalapi;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class InternalApiApplication {
+
+    public static void main(String[] args) {
+        SpringApplication.run(InternalApiApplication.class, args);
+    }
+
+}
