+++ shilly-shally/settings.gradle
@@ -1,1 +1,4 @@
 rootProject.name = 'shilly-shally'
+include 'core-domain'
+include 'external-api'
+include 'internal-api'
