--- shilly-shally/src/main/java/com/shillyshally/ShillyShallyApplication.java
@@ -1,13 +1,0 @@
-package com.shillyshally;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class ShillyShallyApplication {
-
-    public static void main(String[] args) {
-        SpringApplication.run(ShillyShallyApplication.class, args);
-    }
-
-}
