--- shilly-shally/src/main/java/com/shillyshally/application/dto/RestaurantResponse.java
@@ -1,8 +1,0 @@
-package com.shillyshally.application.dto;
-
-public record RestaurantResponse(
-        long id,
-        String name
-) {
-
-}
