--- shilly-shally/src/main/java/com/shillyshally/config/WebConfig.java
@@ -1,21 +1,0 @@
-package com.shillyshally.config;
-
-
-import org.springframework.context.annotation.Configuration;
-import org.springframework.web.servlet.config.annotation.CorsRegistry;
-import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
-
-@Configuration
-public class WebConfig implements WebMvcConfigurer {
-
-    private static final String ALLOWED_METHOD_NAMES = "GET,POST,PUT,DELETE,OPTIONS,PATCH";
-
-    @Override
-    public void addCorsMappings(final CorsRegistry registry) {
-        registry.addMapping("/**")
-                .allowedOrigins("*")
-                .allowedMethods(ALLOWED_METHOD_NAMES.split(","))
-                .exposedHeaders("*");
-    }
-}
-
