--- shilly-shally/src/main/java/com/shillyshally/domain/Restaurant.java
@@ -1,40 +1,0 @@
-package com.shillyshally.domain;
-
-import jakarta.persistence.Entity;
-import jakarta.persistence.EnumType;
-import jakarta.persistence.Enumerated;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
-import lombok.AccessLevel;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Getter
-@Entity
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-@AllArgsConstructor
-@Builder
-public class Restaurant {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    private String name;
-
-    @Enumerated(value = EnumType.STRING)
-    private Category category;
-
-    private String categoryDetail;
-
-    private String phone;
-
-    private String url;
-
-    private Double longitude;
-
-    private Double latitude;
-}
