--- shilly-shally/src/main/java/com/shillyshally/exception/dto/ErrorResponse.java
@@ -1,4 +1,0 @@
-package com.shillyshally.exception.dto;
-
-public record ErrorResponse(String message) {
-}
