--- shilly-shally/src/test/java/com/shillyshally/ShillyShallyApplicationTests.java
@@ -1,13 +1,0 @@
-package com.shillyshally;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class ShillyShallyApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
